Move LinkSpeed to `phy` because it's likely that more 10BaseT/100BaseT PHY will be supported
