+++ src/phy/mod.rs
@@ -1,7 +1,23 @@
+use num_enum::{IntoPrimitive, TryFromPrimitive};
+
 #[cfg(feature = "phy-lan87xxa")]
 mod lan87xxa;
 #[cfg(feature = "phy-lan87xxa")]
 pub use lan87xxa::*;
+
+/// The link speeds supported by this PHY
+#[derive(Clone, Copy, Debug, IntoPrimitive, TryFromPrimitive)]
+#[repr(u8)]
+pub enum LinkSpeed {
+    /// 10BaseT - Half duplex
+    BaseT10HalfDuplex = 0b001,
+    /// 10BaseT - Full duplex
+    BaseT10FullDuplex = 0b101,
+    /// 100BaseT - Half duplex
+    BaseT100HalfDuplex = 0b010,
+    /// 100BaseT - Full duplex
+    BaseT100FullDuplex = 0b110,
+}
 
 pub trait Phy {
     type LinkSpeed;
@@ -18,7 +34,7 @@ pub trait Phy {
 
     /// Poll the link status of this PHY. If it returns `true`, the
     /// link is up. If it returns `false`, the link is down.
-    /// 
+    ///
     // TODO: return an associated type with link information?
     fn poll_link(&mut self) -> bool;
